feat(stage): create an abstraction for FileLocation
