+++ setup.cfg
@@ -70,6 +70,7 @@ sqlalchemy =
     *.typed
     *.sql
     *.csv
+    *.ipynb
 
 [tool:pytest]
 markers =
